Save scores to files
File I/O additions were made to both the GameControl and LevelControl
scripts of Ary's latest commit "MMM". It should be just a matter of
copying these scripts and replacing the ones in your projects to allow
for saving of scores to files, back up as usual. The three text files
provided are examples of output you should expect, they will appear in
the root directory of your project folder.