Scene organisation
Cleaned up scenes with 'folders'... note: if a game object is set to
DontDestroyOnLoad() it will still be destroyed on load if it is a child
of an object which is destroyed on load. So make sure the Game control
object in Neuromend First is never a child of a destructable.